requirements.txt
# Core
torch>=2.2.0
transformers>=4.44.0
datasets>=2.19.0
tokenizers>=0.19.0
sentencepiece>=0.1.99
protobuf<5
tiktoken>=0.7.0

# Training / utilities
numpy>=1.24.0
tqdm>=4.66.0
pyyaml>=6.0.1